fix: add firebase service account key to env variable (#204)
* fix: add firebase service account key to env variable

Prior to this change, a new developer starting server would get an error message saying private key is missing. Installed dotenv to handle environment variables in the code. Updated README with instructions on how to get a private key from firebase and how to add it to the local env file.

* docs: Add instructions for setting up Firebase Service Account